User checking and functional login.

+++ js/include.js
@@ -1,4 +1,11 @@
 $(document).ready(function() {
+  if (localStorage['emails'] === undefined) {
+    localStorage['emails'] = JSON.stringify(["admccarthy@gmail.com", "rsward@gmail.com"]);
+    var rebecca = {"first_name":"Rebecca", "last_name": "Ward", "email": "rsward@gmail.com", "password": "passward", "credit_card": "3734123412341234", "provider": "MasterCard"};
+    var arya = {"first_name":"Arya", "last_name": "McCarthy", "email": "admccarthy@gmail.com", "password": "facebook", "credit_card": "5432432143214321", "provider": "visa"};
+    localStorage['users'] = JSON.stringify([rebecca, arya]);
+  }
+
   var ticket = document.getElementById('active_ticket');
   $("#custom_order_link").click(function(event){
     event.preventDefault();
@@ -23,21 +30,51 @@ $(document).ready(function() {
   // TODO: Make not a cheap hack.
   $(document).on('submit', '#login_form', function (event) {
     event.preventDefault();
-    var first_name = $("#login_email").val(); // CHANGE THIS.
-    $("#user_hello").text(first_name);
-    $("#login_section").hide();
-    $("#user_section").show();
+    var email = $("#login_email").val()
+    if (userExists(email)) {
+      var users = JSON.parse(localStorage['users']);
+      for (var i = 0; i < users.length; ++i) {
+        if (users[i]['email'] === email && users[i]['password'] === $("#login_password").val()) {
+          localStorage.setItem("first_name", users[i]['first_name']);
+          localStorage.setItem("last_name", users[i]['last_name']);
+          localStorage.setItem("email", users[i]['email']);
+          localStorage.setItem("password", users[i]['password']);
+          localStorage.setItem("credit_card", users[i]['credit_card']);
+          localStorage.setItem("provider", users[i]['provider']);
+          showUser();
+          return;
+        }
+      }
+      alert("Incorrect password");
+    } else {
+      alert("Email does not exist.");
+    }
   });
 
   $(document).on('submit', '#login', function (event) {
     event.preventDefault();
-    localStorage.setItem("first_name", $("#first_name").val());
-    localStorage.setItem("last_name", $("#last_name").val());
-    localStorage.setItem("email", $("#email").val());
-    localStorage.setItem("password", $("#password").val());
-    localStorage.setItem("credit_card", $("#credit_card").val());
-    localStorage.setItem("provider", $("#provider").val());
-    window.location = 'index.html';
+    if (!userExists($("#email").val())) {
+      localStorage.setItem("first_name", $("#first_name").val());
+      localStorage.setItem("last_name", $("#last_name").val());
+      localStorage.setItem("email", $("#email").val());
+      localStorage.setItem("password", $("#password").val());
+      localStorage.setItem("credit_card", $("#credit_card").val());
+      localStorage.setItem("provider", $("#provider").val());
+      localStorage['emails'] = appendElementToJsonArray($("#email").val(), localStorage['emails']);
+      var newUser = {
+        "first_name": localStorage['first_name'],
+        "last_name": localStorage['last_name'],
+        "email": localStorage['email'],
+        "password": localStorage['password'],
+        "credit_card": localStorage['credit_card'],
+        "provider": localStorage['provider']
+      };
+      localStorage['users'] = appendElementToJsonArray(newUser, localStorage['users']);
+      window.location = 'index.html';
+    }
+    else {
+      alert("Email is already in use.");
+    }
   });
   showUser();
   $(document).on('submit', '#payment', function(event) {
@@ -47,6 +84,21 @@ $(document).ready(function() {
   $('[name="sides"]').on('change', limitSelectionToOne);
   $('[name="cheese"]').on('change', limitSelectionToOne);
 });
+
+function appendElementToJsonArray(str, json) {
+  console.log(str);
+  var arr = JSON.parse(json);
+  console.log(arr);
+  arr.push(str);
+  console.log(arr);
+  new_json = JSON.stringify(arr);
+  console.log(new_json);
+  return new_json;
+}
+
+function userExists(email) {
+  return (localStorage['emails'] !== undefined) && (JSON.parse(localStorage['emails']).indexOf(email) !== -1);
+}
 
 function limitSelectionToOne(event) {
   var name = event.target.name;
